refactor : move model types to each files
