+++ src/__utils__/logger.ts
@@ -1,0 +1,2 @@
+// TODO: implement logger
+// https://www.fastify.io/docs/latest/Reference/Logging/
