+++ src/models/History.ts
@@ -1,0 +1,6 @@
+export type History = {
+    id: string;
+    repoId?: string;
+    date?: string;
+    state?: string;
+};
