+++ src/models/Repository.ts
@@ -1,0 +1,9 @@
+export type Repository = {
+    id: string;
+    userId?: string;
+    repoName: string;
+    repoUrl: string;
+    repoBranch?: string;
+    deployRepo?: string;
+    state: boolean;
+};
