+++ src/models/User.ts
@@ -1,0 +1,6 @@
+export type User = {
+    userId: string;
+    nickname: string;
+    email?: string;
+    apiKey?: string;
+};
